LPD-46182 Move to below component as we don't need it in the parent and add input fragments
